test: share the fixtures the suites duplicated
The two storage suites carried the same thirty lines of localStorage
stub verbatim, the journey and checkin builders were repeated between
two more, and three suites each wrote their own float comparison.

A fixture that exists in two places drifts: a builder gains a field
on one side only, and the suite on the other keeps asserting against
a shape the type no longer describes. Keeping them in one place under
src/test also gives the next suite something to reach for instead of
copying whichever neighbour looks closest.
