+++ src/utils/journeyPauseStorage.test.ts
@@ -1,36 +1,11 @@
 import assert from 'node:assert/strict';
 import { beforeEach, describe, it } from 'node:test';
+import {
+  installLocalStorage,
+  memoryStorage,
+  throwingStorage
+} from '../test/localStorage.ts';
 import { deletePaused, loadPaused, savePaused } from './journeyPauseStorage.ts';
-
-type StorageStub = Pick<Storage, 'getItem' | 'setItem' | 'removeItem'>;
-
-function installLocalStorage(stub: StorageStub): void {
-  (globalThis as unknown as { window: { localStorage: StorageStub } }).window =
-    { localStorage: stub };
-}
-
-function memoryStorage(): StorageStub & { store: Map<string, string> } {
-  const store = new Map<string, string>();
-
-  return {
-    store,
-    getItem: (key) => store.get(key) ?? null,
-    setItem: (key, value) => {
-      store.set(key, value);
-    },
-    removeItem: (key) => {
-      store.delete(key);
-    }
-  };
-}
-
-function throwingStorage(): StorageStub {
-  const denied = () => {
-    throw new Error('denied');
-  };
-
-  return { getItem: denied, setItem: denied, removeItem: denied };
-}
 
 describe('journeyPauseStorage', () => {
   let storage: ReturnType<typeof memoryStorage>;
